requirements.txt
functions-framework==3.*
python-telegram-bot==20.*
google-cloud-secret-manager==2.* 